Rename stream.py -> core.py
